Suite solveur (segfault...)

+++ include/solveur.h
@@ -16,7 +16,7 @@ tree_game get_child(tree_game t, int i);
 void add_child(tree_game t, game g);
 bool has_child(tree_game t);
 void set_dir_prev(game g, int ind, int dir);
-game sub_solve(game* t, game g, int ind, int ind_tab);
+game sub_solve(game* t, game g, unsigned long ind, unsigned long ind_tab);
 int get_nb_children(tree_game t);
 int solve(game g);
